chores: handling not found

+++ src/controllers/banner.controller.ts
@@ -2,7 +2,7 @@ import { Response } from "express";
 import { IPaginationQuery, IReqUser } from "../utils/interfaces";
 import response from "../utils/response";
 import BannerModel, { bannerDAO, TypeBanner } from "../models/banner.model";
-import { FilterQuery } from "mongoose";
+import { FilterQuery, isValidObjectId } from "mongoose";
 
 
 export default {
@@ -53,6 +53,7 @@ export default {
   findOne: async (req: IReqUser, res: Response) => {
     try {
       const { id } = req.params
+      if (!isValidObjectId(id)) return response.notFound(res, "failed to find a banner")
       const result = await BannerModel.findById(id)
       if (!result) return response.notFound(res, "failed to find a banner")
       response.success(res, result, "success to find a banner")
@@ -63,6 +64,7 @@ export default {
   update: async (req: IReqUser, res: Response) => {
     try {
       const { id } = req.params
+      if (!isValidObjectId(id)) return response.notFound(res, "failed to find a banner")
       const result = await BannerModel.findByIdAndUpdate(id, req.body, { new: true })
       response.success(res, result, "success to update a banner")
     } catch (error) {
@@ -72,6 +74,7 @@ export default {
   remove: async (req: IReqUser, res: Response) => {
     try {
       const { id } = req.params
+      if (!isValidObjectId(id)) return response.notFound(res, "failed to find a banner")
       const result = await BannerModel.findByIdAndDelete(id, { new: true })
       response.success(res, result, "success to remove a banner")
     } catch (error) {
