+++ src/controllers/event.controller.ts
@@ -2,7 +2,7 @@ import { Response } from "express";
 import { IPaginationQuery, IReqUser } from "../utils/interfaces";
 import response from "../utils/response";
 import EventModel, { eventDAO, TEvent } from "../models/event.model";
-import { FilterQuery } from "mongoose";
+import { FilterQuery, isValidObjectId } from "mongoose";
 
 export default {
   create: async (req: IReqUser, res: Response) => {
@@ -52,6 +52,7 @@ export default {
   findOne: async (req: IReqUser, res: Response) => {
     try {
       const { id } = req.params
+      if (!isValidObjectId(id)) return response.notFound(res, "failed to find an event")
       const result = await EventModel.findById(id)
       if (!result) return response.notFound(res, "failed to find an event")
       response.success(res, result, "success to get an event")
@@ -62,6 +63,7 @@ export default {
   update: async (req: IReqUser, res: Response) => {
     try {
       const { id } = req.params
+      if (!isValidObjectId(id)) return response.notFound(res, "failed to find an event")
       const result = await EventModel.findByIdAndUpdate(id, req.body, { new: true })
       response.success(res, result, "success to update an event")
     } catch (error) {
@@ -71,6 +73,7 @@ export default {
   remove: async (req: IReqUser, res: Response) => {
     try {
       const { id } = req.params
+      if (!isValidObjectId(id)) return response.notFound(res, "failed to find an event")
       const result = await EventModel.findByIdAndDelete(id, { new: true })
       response.success(res, result, "success to delete an event")
     } catch (error) {
